[Test] Inital test setup
Use pyproject.toml config options for pytest scenario. Also start with adding a minimal test to make sure that finding the features
by name when working with annotations in the code itself works.
The test envrionment works as expected. Note that this commit does not contain the function defintions used for this test