Refactor metadata file readout
Needing to checkout the metadata branch comes with a great deal of problems. One of them is, if the git repo is dirty, a clean branch switch will not be possible. Utilising git commands that are seldomly used on a daily basis but help with
evaluating file content on other branches are deemed a better fit for this.